feat!: add error handling and rejections
The prior version of the SDK would silently swallow errors that
occurred returning the JSON response without any indication that
a non-OK status code had been encountered.

We have now updated the SDK to instead throw when error responses
are received with a custom error class being used for type narrowing.

Additionally we have updated the tests and types to be more correct
utilising both type guards and the `satisfies` keyword.