fix!: make @react-email/render an optional peer dependency
Several Resend users have been complaining about unexpectedly seeing
`@react-email/render` and `react` in their builds. After analysis, it
was found that this was happening because `@react-email/render` was
listed as a dependency of resend-node.

To fix this, I decided to make `@react-email/render` an optional peer
dependency. This way, it won’t be automatically installed when
`resend-node` is installed.

No functionality should be affected, as `@react-email/render` is
dynamically imported throughout the codebase. However, this is a
breaking change for users who use the react parameter to send emails
with `resend-node`, as they will now need to install
`@react-email/render themselves`.